Split file reader and file input components
