+++ ui/field-file-input.reel/field-file-input.js
@@ -1,0 +1,3 @@
+var Component = require("montage/ui/component").Component;
+
+exports.FieldFileInput = Component.specialize();
