+++ ui/field-file-reader.info/sample/ui/main.reel/main.html
@@ -16,10 +16,10 @@
             }
        },
         "default": {
-            "prototype": "blue-shark/ui/field-file-upload.reel",
+            "prototype": "blue-shark/ui/field-file-reader.reel",
             "properties": {
                 "element": {"#": "default"},
-                "label": "Field File Upload"
+                "label": "Field File Reader"
             }
         }
     }
