+++ ui/field-file-reader.reel/field-file-reader.html
@@ -1,0 +1,50 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <link rel="stylesheet" type="text/css" href="field-file-reader.css">
+    <script type="text/montage-serialization">
+    {
+        "owner": {
+            "properties": {
+                "element": {"#": "owner"}
+            }
+        },
+        "field": {
+            "prototype": "ui/field.reel",
+            "properties": {
+                "element": {"#": "field"}
+            },
+            "bindings": {
+                "label": {"<-": "@owner.label"},
+                "documentationReference": {"<-": "@owner.documentationReference"},
+                "disabled": {"<-": "@owner.disabled"},
+                "errorMessage": {"<-": "@owner.errorMessage || @control.errorMessage"},
+                "hasError": {"<-": "@owner.hasError || !!@control.errorMessage"},
+                "isMandatory": {"<-": "@owner.isMandatory"}
+            }
+        },
+        "control": {
+            "prototype": "ui/file-reader.reel",
+            "properties": {
+                "element": {"#": "fileReader"}
+            },
+            "bindings": {
+                "file": {"<->": "@owner.file"},
+                "supportedExtensions": {"<-": "@owner.supportedExtensions"},
+                "supportedFileTypes": {"<-": "@owner.supportedFileTypes"},
+                "maxFileSize": {"<-": "@owner.maxFileSize"},
+                "resultType": {"<-": "@owner.resultType"},
+                "data": {"<->": "@owner.data"}
+            }
+        }
+    }
+    </script>
+</head>
+<body>
+    <div data-montage-id="owner" class="FieldFileReader">
+        <div data-montage-id="field">
+            <div data-montage-id="fileReader"></div>
+        </div>
+    </div>
+</body>
+</html>
