+++ ui/field-file-reader.reel/field-file-reader.js
@@ -1,0 +1,3 @@
+var Component = require("montage/ui/component").Component;
+
+exports.FieldFileReader = Component.specialize();
