--- ui/field-file-upload.reel/_field-file-upload.css
@@ -1,3 +1,0 @@
-.FieldFileUpload {
-
-}
