--- ui/field-file-upload.reel/field-file-upload.js
@@ -1,10 +1,0 @@
-/**
- * @module ui/field-file-upload.reel
- */
-var Component = require("montage/ui/component").Component;
-
-/**
- * @class FieldFileUpload
- * @extends Component
- */
-exports.FieldFileUpload = Component.specialize();
