+++ ui/file-input.info/sample/ui/main.reel/main.html
@@ -1,0 +1,45 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <link rel="stylesheet" type="text/css" href="main.css">
+    <script type="text/montage-serialization">
+    {
+        "owner": {
+            "properties": {
+                "element": {"#": "owner"}
+            }
+        },
+
+        "icons": {
+            "prototype": "blue-shark/ui/icons.reel",
+            "properties": {
+                "element": {"#": "icons"}
+            }
+        },
+
+        "fileInput": {
+            "prototype": "blue-shark/ui/file-input.reel",
+            "properties": {
+                "element": {"#": "fileInput"}
+            }
+        },
+
+        "fileInput2": {
+            "prototype": "blue-shark/ui/file-input.reel",
+            "properties": {
+                "element": {"#": "fileInput2"}
+            }
+        }
+    }
+    </script>
+</head>
+<body>
+    <div data-montage-id="owner" class="Main">
+        <div data-montage-id="icons"></div>
+        <div class="container">
+            <div data-montage-id="fileInput"></div>
+            <div data-montage-id="fileInput2"></div>
+        </div>
+    </div>
+</body>
+</html>
