+++ ui/file-reader.info/sample/package.json
@@ -1,0 +1,14 @@
+{
+    "name": "file-upload-sample",
+    "version": "0.1.0",
+    "dependencies": {
+        "blue-shark": "*",
+        "montage": "*"
+    },
+        "mappings": {
+        "montage": "../../../node_modules/montage",
+        "blue-shark": "../../../"
+    }
+}
+
+
