+++ ui/file-reader.info/sample/ui/main.reel/main.html
@@ -17,17 +17,17 @@
             }
         },
 
-        "fileUpload": {
-            "prototype": "blue-shark/ui/file-upload.reel",
+        "fileReader": {
+            "prototype": "blue-shark/ui/file-reader.reel",
             "properties": {
-                "element": {"#": "fileUpload"}
+                "element": {"#": "fileReader"}
             }
         },
 
-        "fileUpload2": {
-            "prototype": "blue-shark/ui/file-upload.reel",
+        "fileReader2": {
+            "prototype": "blue-shark/ui/file-reader.reel",
             "properties": {
-                "element": {"#": "fileUpload2"}
+                "element": {"#": "fileReader2"}
             }
         }
     }
@@ -37,8 +37,8 @@
     <div data-montage-id="owner" class="Main">
         <div data-montage-id="icons"></div>
         <div class="container">
-            <div data-montage-id="fileUpload"></div>
-            <div data-montage-id="fileUpload2"></div>
+            <div data-montage-id="fileReader"></div>
+            <div data-montage-id="fileReader2"></div>
         </div>
     </div>
 </body>
