+++ ui/file-reader.info/sample/ui/main.reel/main.js
@@ -1,0 +1,10 @@
+/**
+ * @module ui/main.reel
+ */
+var Component = require("montage/ui/component").Component;
+
+/**
+ * @class Main
+ * @extends Component
+ */
+exports.Main = Component.specialize();
