--- ui/file-upload.reel/file-upload.css
@@ -1,1 +1,0 @@
-.FileUpload{position:relative;display:-webkit-box;display:-ms-flexbox;display:flex;-ms-flex-wrap:wrap;flex-wrap:wrap;-webkit-box-align:center;-ms-flex-align:center;align-items:center}.FileUpload.has-progress{padding-bottom:1.5em}.FileUpload-path{font-family:lato;font-weight:200;font-size:.75em}.FileUpload-button{box-sizing:border-box;font-family:Lato,Helvetica,Arial,sans-serif;font-size:.75em;font-weight:300;min-width:6em;padding:0 1em;text-align:center;color:#e6e6e6;background:#fff;border-radius:.3em;border:0;margin:0;cursor:pointer;-webkit-appearance:none;-moz-appearance:none;appearance:none;text-shadow:0 1px 0 rgba(0,0,0,.5);background-color:#666;background-image:-webkit-linear-gradient(transparent,rgba(0,0,0,.2));background-image:linear-gradient(transparent,rgba(0,0,0,.2));box-shadow:inset 0 1px hsla(0,0%,100%,.15),inset 0 -1px rgba(0,0,0,.2),0 1px 2px rgba(0,0,0,.4);-webkit-transition:all .25s ease-in-out;transition:all .25s ease-in-out;display:inline-block;vertical-align:middle;line-height:2.5em;margin-right:16px;margin-right:1rem}.FileUpload-button:hover{background-image:-webkit-linear-gradient(transparent,transparent);background-image:linear-gradient(transparent,transparent);box-shadow:inset 0 1px hsla(0,0%,100%,.15),inset 0 -1px rgba(0,0,0,.3),0 2px 6px 1px rgba(0,0,0,.4)}.FileUpload-button:active{background-image:-webkit-linear-gradient(rgba(0,0,0,.2),transparent);background-image:linear-gradient(rgba(0,0,0,.2),transparent);box-shadow:inset 0 -1px hsla(0,0%,100%,.1),inset 0 1px rgba(0,0,0,.3);outline:none;text-shadow:none}.FileUpload-button:focus{outline:none;box-shadow:inset 0 0 0 1px #2089d3}.FileUpload-button.montage--disabled{box-shadow:inset 0 1px hsla(0,0%,100%,0),inset 0 -1px transparent,0 1px 2px transparent;background-image:-webkit-linear-gradient(transparent,transparent);background-image:linear-gradient(transparent,transparent);text-shadow:0 1px 0 transparent;cursor:default;cursor:not-allowed}.FileUpload-button.montage--disabled:not(.is-processing){opacity:.5;background-color:#666}.FileUpload-input{position:absolute;left:-9999px}.FileUpload-progress{display:none;position:absolute;bottom:0;left:0;right:0}.FileUpload.has-progress .FileUpload-progress{display:-webkit-box;display:-ms-flexbox;display:flex}
