+++ ui/main.reel/main.html
@@ -31,7 +31,8 @@
                 <li><a href="ui/datetime.info/sample/">Date/Time</a></li>
                 <li><a href="ui/duration.info/sample/">Duration</a></li>
                 <li><a href="ui/field.info/sample/">Field</a></li>
-                <li><a href="ui/file-upload.info/sample/">File Upload</a></li>
+                <li><a href="ui/file-input.info/sample/">File Input</a></li>
+                <li><a href="ui/file-reader.info/sample/">File Reader</a></li>
                 <li><a href="ui/multiple-select-grid.info/sample/">Multiple Select Grid</a></li>
                 <li><a href="ui/multiple-select.info/sample/">Multiple Select</a></li>
                 <li><a href="ui/number-input.info/sample/">Input - Number</a></li>
